Add files via upload
First step in creating a bot for the online game RuneScape.  This is used to capture the game window, take screenshots to use for cascade training, and draw rectangles around the specified object in game.  This project is still a WIP as I play with the data for the positive and negative files to create a solution that does have a large amount of false positives.